Remove discussion hero, put title at the top of the post stream
