+++ less/forum/DiscussionHeader.less
@@ -1,11 +1,14 @@
-.DiscussionHero {
+.DiscussionHeader {
+  margin: 30px 0;
+  text-align: center;
+
   .badges {
     margin-right: 10px;
     margin-bottom: -2px;
     vertical-align: 2px;
   }
 }
-.DiscussionHero-items {
+.DiscussionHeader-items {
   padding: 0;
   margin: 0;
   list-style: none;
@@ -14,7 +17,9 @@
     display: inline-block;
   }
 }
-.DiscussionHero-title {
+.DiscussionHeader-title {
   display: inline;
   vertical-align: middle;
+  font-size: 22px;
+  font-weight: normal;
 }
