+++ less/forum/app.less
@@ -3,7 +3,7 @@
 @import "ActivityPage.less";
 @import "AvatarEditor.less";
 @import "Composer.less";
-@import "DiscussionHero.less";
+@import "DiscussionHeader.less";
 @import "DiscussionList.less";
 @import "DiscussionListItem.less";
 @import "DiscussionPage.less";
